+++ src/studio_lib/property_grid.cpp
@@ -13,6 +13,7 @@
 #include "terrain_editor.h"
 #include "utils.h"
 #include <cmath>
+#include <cstdlib>
 
 
 const char* PropertyGrid::getComponentTypeName(Lumix::ComponentUID cmp) const
@@ -168,7 +169,7 @@ void PropertyGrid::showProperty(Lumix::IPropertyDescriptor& desc, int index, Lum
 		Lumix::copyString(buf, (const char*)stream.getData());
 		auto& resource_descriptor = dynamic_cast<Lumix::ResourcePropertyDescriptorBase&>(desc);
 		auto rm_type = resource_descriptor.getResourceType();
-		auto asset_type = m_asset_browser.getTypeFromResourceManagerType(rm_type);
+		auto asset_type = m_asset_browser.getTypeIndexFromManagerType(rm_type);
 		if (m_asset_browser.resourceInput(desc.getName(),
 				StringBuilder<20>("", (Lumix::uint64)&desc),
 				buf,
